+++ packages/mistralai-gcp/src/models/components/index.ts
@@ -18,6 +18,7 @@ export * from "./fimcompletionstreamrequest.js";
 export * from "./function.js";
 export * from "./functioncall.js";
 export * from "./functionname.js";
+export * from "./prediction.js";
 export * from "./referencechunk.js";
 export * from "./responseformat.js";
 export * from "./responseformats.js";